pyproject.toml
[project]
name = "claude-brain"
version = "2.0.0"
description = "Agente de desarrollo autónomo basado en Claude"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "supabase>=2.3.0",
    "httpx>=0.27.0",
    "mem0ai>=0.1.0",
    "pyyaml>=6.0.0",
    "websockets>=12.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
browser = ["playwright>=1.42.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
